IRContext cleared
